// extensions/discord/src/send.types.ts
export type DiscordRuntimeAccountContext = {
  cfg: AutopusConfig;
  accountId: string;
};

export type DiscordReactOpts = {
  cfg: AutopusConfig;
  accountId?: string;
  token?: string;
  rest?: RequestClient;
  verbose?: boolean;
  retry?: RetryConfig;
};

export type DiscordReactionRuntimeContext = DiscordRuntimeAccountContext & {
  rest: RequestClient;
};